Revert "fixed endless loop when the channel file is missing"
This reverts commit baf7ceaa6d832bc4557f59aca7a12e80e46ecebe which
caused the import to fail after processing the first data file.
Closes #32

+++ src/EPGImport/EPGImport.py
@@ -216,12 +216,13 @@ class EPGImport:
 
 	def afterChannelDownload(self, result, filename, deleteFile=True):
 		print>>log, "[EPGImport] afterChannelDownload", filename
-		if filename is None:
-			self.channelDownloadFail(Exception("File does not exist"));
-			return
-		elif not os.path.getsize(filename):
-			self.channelDownloadFail(Exception('File is empty'))
-			return
+		if filename:
+			try:
+				if not os.path.getsize(filename):
+					raise Exception, "File is empty"
+			except Exception, e:
+				self.channelDownloadFail(e)
+				return
 		if twisted.python.runtime.platform.supportsThreads():
 			print>>log, "[EPGImport] Using twisted thread"
 			threads.deferToThread(self.doThreadRead, filename).addCallback(lambda ignore: self.nextImport())
@@ -229,7 +230,7 @@ class EPGImport:
 		else:
 			self.iterator = self.createIterator(filename)
 			reactor.addReader(self)
-		if deleteFile:
+		if deleteFile and filename:
 			try:
 				os.unlink(filename)
 			except Exception, e:
